renames all-terms page to glossary
