+++ public/locales/en/common.json
@@ -1,7 +1,7 @@
 {
     "glossetaTitle": "Glosseta",
     "glossetaNavbarButtonA11yText": "Takes you to the main search page",
-    "allTermsButton": "All Terms",
+    "glossaryButton": "Glossary",
     "searchButtonTitle": "Search",
     "scrollToTheTopButton": "Click here to scroll to the top of the page",
