+++ src/types/glossary-item.ts
